enums folder must be enum psr standard for autoloading
