+++ src/APICore.php
@@ -2,9 +2,9 @@
 
 namespace Johnrdorazio\LitCal;
 
-use Johnrdorazio\LitCal\enum\AcceptHeader;
-use Johnrdorazio\LitCal\enum\RequestMethod;
-use Johnrdorazio\LitCal\enum\RequestContentType;
+use Johnrdorazio\LitCal\Enum\AcceptHeader;
+use Johnrdorazio\LitCal\Enum\RequestMethod;
+use Johnrdorazio\LitCal\Enum\RequestContentType;
 
 class APICore
 {
