+++ src/Enum/CalendarType.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Johnrdorazio\LitCal\enum;
+namespace Johnrdorazio\LitCal\Enum;
 
 class CalendarType
 {
