+++ src/Enum/CorpusChristi.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace Johnrdorazio\LitCal\enum;
+namespace Johnrdorazio\LitCal\Enum;
 
-class Ascension
+class CorpusChristi
 {
-    public const THURSDAY       = "THURSDAY";
-    public const SUNDAY         = "SUNDAY";
+    public const THURSDAY          = "THURSDAY";
+    public const SUNDAY            = "SUNDAY";
     public static array $values = [ "THURSDAY", "SUNDAY" ];
 
     public static function isValid($value)
